Extend database API to support guilds

+++ src/MHServerEmu.DatabaseAccess/Models/DBGuildMember.cs
@@ -4,9 +4,9 @@
     {
         public long PlayerDbGuid { get; set; }
         public long GuildId { get; set; }
-        public int Membership { get; set; }
+        public long Membership { get; set; }    // This needs to be long for our Dapper/System.Data.SQLite combo.
 
-        public DBGuildMember(long playerDbGuid, long guildId, int membership)
+        public DBGuildMember(long playerDbGuid, long guildId, long membership)
         {
             PlayerDbGuid = playerDbGuid;
             GuildId = guildId;
